JSON bug

+++ src/main/java/net/eithon/plugin/donationboard/logic/PlayerInfo.java
@@ -71,8 +71,8 @@ public class PlayerInfo implements IJson<PlayerInfo>, IUuidAndName  {
 	public PlayerInfo fromJson(Object json) {
 		JSONObject jsonObject = (JSONObject) json;
 		this._eithonPlayer = EithonPlayer.getFromJSon(jsonObject.get("player"));
-		this._remainingDonationTokens = (int) jsonObject.get("remainingDonationTokens");
-		this._totalTokensDonated = (long) jsonObject.get("totalTokensDonated");
+		this._remainingDonationTokens = ((Long) jsonObject.get("remainingDonationTokens")).intValue();
+		this._totalTokensDonated = ((Long) jsonObject.get("totalTokensDonated")).intValue();
 		this._totalMoneyDonated = (double) jsonObject.get("totalMoneyDonated");
 		return this;
 	}
